# api-spec_openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Adversarial Candidate Ranking System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ranking/start:
    post:
      operationId: startRanking
      tags: [ranking]
      summary: Start a ranking job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RankingStartInput"
      responses:
        "201":
          description: Ranking job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingJob"
...
  /ranking/status/{jobId}:
    get:
      operationId: getRankingStatus
      tags: [ranking]
      summary: Get ranking job status
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ranking job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingJob"
  /ranking/results/{jobId}:
    get:
      operationId: getRankingResults
      tags: [ranking]
      summary: Get complete ranking results
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ranking results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingResults"

  # ADVERSARY
...
components:
  schemas:
...
    RankingStartInput:
      type: object
      required: [jdId, candidateIds]
      properties:
        jdId:
          type: integer
        candidateIds:
          type: array
          items:
            type: integer
    RankingJob:
      type: object
      required: [id, status, totalCandidates, progress, createdAt]
      properties:
        id:
          type: integer
        jdId:
          type: integer
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress:
          type: integer
        totalCandidates:
          type: integer
        jobTitle:
          type: ["string", "null"]
        createdAt:
          type: string
    RankedCandidate:
      type: object
      required: [candidateId, name, finalScore, fitScore, robustnessScore, stabilityScore, ranking, isRobust, redFlags, explanation, recommendation]
      properties:
        candidateId:
          type: integer
        name:
          type: string
        finalScore:
          type: number
        fitScore:
          type: number
        robustnessScore:
          type: number
        stabilityScore:
          type: number
        ranking:
          type: integer
        isRobust:
          type: boolean
        manipulationRisk:
          type: string
          enum: [low, medium, high]
        redFlags:
          type: array
          items:
            $ref: "#/components/schemas/RedFlag"
        skillsMatch:
          $ref: "#/components/schemas/SkillsMatch"
        explanation:
          type: string
        recommendation:
          type: string
          enum: [strong_hire, hire, consider, reject]
        strengths:
          type: array
          items:
            type: string
        gaps:
          type: array
          items:
            type: string
        adversarialTests:
          type: array
          items:
            $ref: "#/components/schemas/AdversarialTest"
    RedFlag:
      type: object
      required: [type, severity, details]
      properties:
        type:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        details:
          type: string
    SkillsMatch:
      type: object
      required: [matched, missing, extra]
      properties:
        matched:
          type: array
          items:
            type: string
        missing:
          type: array
          items:
            type: string
        extra:
          type: array
          items:
            type: string
    AdversarialTest:
      type: object
      required: [perturbationType, originalRank, perturbedRank, positionChange]
      properties:
        perturbationType:
          type: string
        originalRank:
          type: integer
        perturbedRank:
          type: integer
        positionChange:
          type: integer
    RankingSummary:
      type: object
      required: [totalCandidates, flaggedForManipulation, highlyRobust, averageRobustness, timeTaken]
      properties:
        totalCandidates:
          type: integer
        flaggedForManipulation:
          type: integer
        highlyRobust:
          type: integer
        averageRobustness:
          type: number
        timeTaken:
          type: string
    RankingResults:
      type: object
      required: [jobId, rankedCandidates, summary]
      properties:
        jobId:
          type: integer
        rankedCandidates:
          type: array
          items:
            $ref: "#/components/schemas/RankedCandidate"
        summary:
          $ref: "#/components/schemas/RankingSummary"

    # ADVERSARY
...